child process success, fix scmp & possibly bloomberg tmr
add outline.com list.
